Move app testing wrappers to chip.testing.apps
