+++ src/python_testing/TC_ECOINFO_2_2.py
@@ -45,7 +45,7 @@ import tempfile
 
 import chip.clusters as Clusters
 from chip.interaction_model import Status
-from chip.testing.tasks import AppServerSubprocess
+from chip.testing.apps import AppServerSubprocess
 from matter_testing_support import MatterBaseTest, TestStep, async_test_body, default_matter_test_main
 from mobly import asserts
