+++ src/python_testing/matter_testing_infrastructure/chip/testing/apps.py
@@ -1,0 +1,69 @@
+# Copyright (c) 2024 Project CHIP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import os
+import signal
+import tempfile
+
+from .tasks import Subprocess
+
+
+class AppServerSubprocess(Subprocess):
+    """Wrapper class for starting an application server in a subprocess."""
+
+    # Prefix for log messages from the application server.
+    PREFIX = b"[SERVER]"
+
+    def __init__(self, app: str, storage_dir: str, discriminator: int,
+                 passcode: int, port: int = 5540):
+        self.kvs_fd, kvs_path = tempfile.mkstemp(dir=storage_dir, prefix="kvs-app-")
+        # Start the server application with dedicated KVS storage.
+        super().__init__(app, "--KVS", kvs_path,
+                         '--secured-device-port', str(port),
+                         "--discriminator", str(discriminator),
+                         "--passcode", str(passcode),
+                         output_cb=lambda line, is_stderr: self.PREFIX + line)
+
+    def __del__(self):
+        # Do not leak KVS file descriptor.
+        os.close(self.kvs_fd)
+
+
+class IcdAppServerSubprocess(AppServerSubprocess):
+    """Wrapper class for starting an ICD application server in a subprocess."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.paused = False
+
+    def pause(self, check_state: bool = True):
+        if check_state and self.paused:
+            raise ValueError("ICD TH Server unexpectedly is already paused")
+        if not self.paused:
+            # Stop (halt) the ICD server process by sending a SIGTOP signal.
+            self.p.send_signal(signal.SIGSTOP)
+            self.paused = True
+
+    def resume(self, check_state: bool = True):
+        if check_state and not self.paused:
+            raise ValueError("ICD TH Server unexpectedly is already running")
+        if self.paused:
+            # Resume (continue) the ICD server process by sending a SIGCONT signal.
+            self.p.send_signal(signal.SIGCONT)
+            self.paused = False
+
+    def terminate(self):
+        # Make sure the ICD server process is not paused before terminating it.
+        self.resume(check_state=False)
+        super().terminate()
